third_party/blink/web_tests/fast/block/positioning/offsetLeft-offsetTop-borders-expected.txt

jee

-------------------------------------------------offsetLeft:12
-------------------------------------------------offsetLeft:10
-------------------------------------------------offsetLeft:8
-------------------------------------------------offsetLeft:0
-------------------------------------------------offsetTop:12
-------------------------------------------------offsetTop:10
-------------------------------------------------offsetTop:8
-------------------------------------------------offsetTop:0
